Optimize and finish the medial axis generation
Now the selected objects in Rhino (whatever format, e.g., brep, mesh)
can be converted into OFF file with the help of the convertor embedded
in meshlab. The generated model has been preprocessed by merging close
vertices. The conversion is run in the back end and the speed has been
elevated. The only buge and limitation is: the user has to select the
object first and then hit the 'GENERATE' button and geometries with
sharp edges may reduce the accuracy of the generated medial axis (this
is also the limitation of the medial axis generation algorithm).